Create shipping.py
This exercise is to ask for user input based on the weight of the product they intend to ship. The company will evaluate the best prices 
given from three different shipping options, puts the values calculated onto a list, and then outputs the cheapest option. This exercise is completed in 
the Visual Studio Code IDE in Python 3.0.